update theming to accept named themes
